+++ core/main_method/solver.py
@@ -8,6 +8,7 @@ __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
 
+import traceback
 from typing import (
     Tuple,
     List,
@@ -30,7 +31,12 @@ from core.libs import (
 
 
 def resolve(self):
-    """Resolve: Use Solvespace lib."""
+    """Resolve: Using three libraries to solve the system.
+    
+    + Pyslvs
+    + Python-Solvespace
+    + Sketch Solve
+    """
     vpoints = self.EntitiesPoint.dataTuple()
     solve_kernel = self.planarsolver_option.currentIndex()
     try:
@@ -39,12 +45,12 @@ def resolve(self):
                 self.getTriangle(),
                 {n: 'P{}'.format(n) for n in range(len(vpoints))},
                 vpoints,
-                [v[-1] for v in self.InputsWidget.getInputsVariables()]
+                tuple(v[-1] for v in self.InputsWidget.inputPair())
             )
         elif solve_kernel == 1:
             result, _ = slvsProcess(
                 vpoints,
-                tuple(self.InputsWidget.getInputsVariables())
+                tuple(self.InputsWidget.inputPair())
                 if not self.freemode_button.isChecked() else ()
             )
         elif solve_kernel == 2:
@@ -54,7 +60,7 @@ def resolve(self):
             )
     except Exception as e:
         if self.consoleerror_option.isChecked():
-            print(e)
+            print(traceback.format_exc())
         self.ConflictGuide.setToolTip(str(e))
         self.ConflictGuide.setStatusTip("Error: {}".format(e))
         self.ConflictGuide.setVisible(True)
@@ -211,25 +217,29 @@ def getCollection(self) -> Dict[str, Union[
     }
 
 
-def getTriangle(self,
-    vpoints: Optional[Tuple[VPoint]] = None
-) -> List[Tuple[str]]:
+def getTriangle(self, vpoints: Optional[Tuple[VPoint]] = None) -> List[Tuple[str]]:
     """Update triangle expression here.
 
     Special function for VPoints.
     """
     if vpoints is None:
         vpoints = self.EntitiesPoint.dataTuple()
+    status = {}
     exprs = vpoints_configure(
         vpoints,
-        tuple((b, d) for b, d, a in self.InputsWidget.inputPair())
+        tuple((b, d) for b, d, a in self.InputsWidget.inputPair()),
+        status
     )
     data_dict, _ = data_collecting(
         exprs,
         {n: 'P{}'.format(n) for n in range(len(vpoints))},
         vpoints
     )
-    self.EntitiesExpr.setExpr(exprs, data_dict)
+    self.EntitiesExpr.setExpr(
+        exprs,
+        data_dict,
+        tuple(p for p, s in status.items() if not s)
+    )
     return exprs