+++ changelog.md
@@ -1,6 +1,23 @@
 Change Log
 ===
 
+Version 18.07
+===
+
+Features
+---
+
++ Dimensional synthesis data settings changed.
++ Input variables has been changed.
++ New solving kernel "Sketch Solve".
+
+Development
+---
+
++ Merge "-w" flag to "-d" flag.
++ New format with dimensional synthesis function.
++ Example list has been move to "pyslvs" core kernel.
+
 Version 18.06
 ===
 
@@ -39,7 +56,7 @@ Features
     + Add color picker in setting interface.
 + Solvespace format:
     + Comments will be generate into a new layout in Solvespace format.
-    + Simple reading function for *.slvs format (only support P joint).
+    + Simple reading function for \*.slvs format (only support R joint).
     + New option to generate part files.
 + Path record:
     + Add "copy path" function in path context menu.